pyproject.toml
[tool.poetry]
name = "termslime"
version = "0.2.0"
description = "Termslime displays images in your terminal with true colors."
license = "MIT"
authors = ["Gary Zhang <garyzhang2002@hotmail.com>"]
maintainers = ["Gary Zhang <garyzhang2002@hotmail.com>"]
readme = "README.md"
homepage = "https://github.com/garyzbm/termslime"
repository = "https://github.com/garyzbm/termslime"
documentation = "https://github.com/garyzbm/termslime"
keywords = ["cli", "image", "terminal"]

[tool.poetry.dependencies]
python = "^3.7"
colorful = "^0.5.5"
pillow = "^9.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tslime = "termslime.__main__:__main"